# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "idp API",
        "contact": {
            "name": "idp Team",
            "email": "idp@overal-x.org"
        },
        "version": "1.0"
    },
    "paths": {
        "/auth/login/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "operationId": "loginUser",
                "parameters": [
                    {
                        "description": "Login User",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.LoginUserArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.LoginUserResult"
                        }
                    }
                }
            }
        },
        "/auth/refresh-access-token/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "operationId": "refreshAccessToken",
                "parameters": [
                    {
                        "description": "Refresh Access Token",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RefreshAccessTokenArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.LoginUserResult"
                        }
                    }
                }
            }
        },
        "/auth/register/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "operationId": "registerUser",
                "parameters": [
                    {
                        "description": "Register User",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RegisterUserArgs"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.RegisterUserResult"
                        }
                    }
                }
            }
        },
...
        "/deployments/deploy": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployments"
                ],
                "operationId": "deployRepo",
                "parameters": [
                    {
                        "description": "Deploy Repo",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.DeployRepoArgs"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/user/profile/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "operationId": "getProfileUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.UserModel"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "operationId": "updateProfileUser",
                "parameters": [
                    {
                        "description": "Update User",
                        "name": "args",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateUserArgs"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/db.UserModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.Currency": {
            "type": "string",
            "enum": [
                "NGN"
            ],
            "x-enum-varnames": [
                "CurrencyNgn"
            ]
        },
        "db.DeploymentLogModel": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "deployment": {
                    "$ref": "#/definitions/db.DeploymentModel"
                },
                "deploymentId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jobId": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.DeploymentModel": {
            "type": "object",
            "properties": {
                "DeploymentLog": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.DeploymentLogModel"
                    }
                },
                "actor": {
                    "type": "string"
                },
                "commitHash": {
                    "type": "string"
                },
                "commitMessage": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machine": {
                    "$ref": "#/definitions/db.MachineModel"
                },
                "machineId": {
                    "type": "string"
                },
                "repoConnection": {
                    "$ref": "#/definitions/db.RepoConnectionModel"
                },
                "repoConnectionId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/db.DeploymentStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.DeploymentStatus": {
            "type": "string",
            "enum": [
                "IDLE",
                "IN_PROGRESS",
                "SUCCESSFUL",
                "FAILED"
            ],
            "x-enum-varnames": [
                "DeploymentStatusIdle",
                "DeploymentStatusInProgress",
                "DeploymentStatusSuccessful",
                "DeploymentStatusFailed"
            ]
        },
        "db.GithubAccountConnectionModel": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "githubEmail": {
                    "type": "string"
                },
                "githubId": {
                    "type": "string"
                },
                "githubInstallationId": {
                    "type": "integer"
                },
                "githubUsername": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/db.UserModel"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "db.MachineModel": {
            "type": "object",
            "properties": {
                "Deployment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.DeploymentModel"
                    }
                },
                "Network": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.NetworkModel"
                    }
                },
                "RepoConnection": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.RepoConnectionModel"
                    }
                },
                "containerId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machineImage": {
                    "type": "string"
                },
                "machineName": {
                    "type": "string"
                },
                "machineStatus": {
                    "$ref": "#/definitions/db.MachineStatus"
                },
                "owner": {
                    "$ref": "#/definitions/db.UserModel"
                },
                "ownerId": {
                    "type": "string"
                },
                "plan": {
                    "$ref": "#/definitions/db.MachinePlanModel"
                },
                "planId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.MachinePlanModel": {
            "type": "object",
            "properties": {
                "Machine": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.MachineModel"
                    }
                },
                "cpu": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "$ref": "#/definitions/db.Currency"
                },
                "deletedAt": {
                    "type": "string"
                },
                "hourlyRate": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                },
                "monthlyRate": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.MachineStatus": {
            "type": "string",
            "enum": [
                "CREATING",
                "RUNNING",
                "REBOOTING",
                "SHUTTING_DOWN",
                "SHUTDOWN"
            ],
            "x-enum-varnames": [
                "MachineStatusCreating",
                "MachineStatusRunning",
                "MachineStatusRebooting",
                "MachineStatusShuttingDown",
                "MachineStatusShutdown"
            ]
        },
        "db.NetworkModel": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "destinationPort": {
                    "type": "integer"
                },
                "hostName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ingressId": {
                    "type": "string"
                },
                "listeningPort": {
                    "type": "integer"
                },
                "machine": {
                    "$ref": "#/definitions/db.MachineModel"
                },
                "machineId": {
                    "type": "string"
                },
                "protocol": {
                    "type": "string"
                },
                "serviceId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.RepoConnectionModel": {
            "type": "object",
            "properties": {
                "Deployment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.DeploymentModel"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machine": {
                    "$ref": "#/definitions/db.MachineModel"
                },
                "machineId": {
                    "type": "string"
                },
                "repoId": {
                    "type": "string"
                },
                "repoName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.UserModel": {
            "type": "object",
            "properties": {
                "GithubAccountConnection": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.GithubAccountConnectionModel"
                    }
                },
                "Machine": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.MachineModel"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "internals.AuthTokens": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "services.DeployRepoArgs": {
            "type": "object",
            "properties": {
                "connectionId": {
                    "type": "string"
                },
                "ref": {
                    "type": "string"
                }
            }
        },
...
        "types.LoginUserArgs": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "types.LoginUserResult": {
            "type": "object",
            "properties": {
                "tokens": {
                    "$ref": "#/definitions/internals.AuthTokens"
                }
            }
        },
        "types.RefreshAccessTokenArgs": {
            "type": "object",
            "required": [
                "refreshToken"
            ],
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "types.RegisterUserArgs": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "types.RegisterUserResult": {
            "type": "object",
            "properties": {
                "GithubAccountConnection": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.GithubAccountConnectionModel"
                    }
                },
                "Machine": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.MachineModel"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tokens": {
                    "$ref": "#/definitions/internals.AuthTokens"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "types.UpdateUserArgs": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}